The beginning of a refactor...
I haven't really done much in the way of programming at home for quite some time, but this will hopefully be the first of at least a handful of changes to get some of these old (and slightly embarrassing) repos in better shape.

First things first. I fixed the markdown syntax of the README.md file...